requirements.txt
newspaper3k>=0.2.8
httpx>=0.27
streamlit>=1.28
crewai>=0.3
openai>=1.4 
pandas>=2.2
requests>=2.31
python-dotenv>=1.0
tenacity>=8.2
lxml_html_clean 
spacy>=3.7
tiktoken>=0.6
faiss-cpu>=1.7.4
numpy>=1.26
folium>=0.15
typing_extensions>=4.7
feedparser>=6.0
tenacity>=8.2
newspaper3k>=0.2.8
playwright>=1.44
beautifulsoup4>=4.12
pytesseract 



